+++ app/ferroehr/src/service/definition/mod.rs
@@ -43,6 +43,7 @@
 
 mod adl14;
 mod adl2;
+mod binding_uri;
 pub(super) mod lineage;
 mod opt14_convert;
 mod query;